# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Hack as a Service API
  version: 1.0.0
servers:
  - url: https://hackclub.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /apps/{slug}/domains:
    get:
      summary: Fetch an app's domains
      tags:
        - Apps
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
          example: dinopoll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
        "500":
          description: Internal server error
        "404":
          description: App not found
        "401":
          description: Unauthorized
    post:
      summary: Add a custom domain to an app
      tags:
        - Apps
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
          example: dinopoll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: string
              required:
                - domain
              example:
                domain: hackclub.com
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
        "500":
          description: Internal server error
        "409":
          description: Conflict
        "404":
          description: App not found
        "401":
          description: Unauthorized
  /domains/{id}/verify:
    post:
      summary: Attempt to verify a domain
      tags:
        - Domains
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          example: 7
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "500":
          description: Internal server error
        "409":
          description: Conflict (domain already verified on another app)
        "404":
          description: Domain not found
        "401":
          description: Unauthorized
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: haas_token
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Domain:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        domain:
          type: string
        verified:
          type: boolean
          readOnly: true
        app_id:
          type: integer
          readOnly: true
      required:
        - id
        - domain
        - verified
        - app_id
      example:
        id: 7
        app_id: 5
        domain: dinopoll.hackclub.app
        verified: true